Update views.py
Changes done after the `XSS report` forwarded yesterday to @juhanurmi

XSS on redirection can be avoided simply checking if the provided `URI-scheme` is valid or not.
In this case (XSS on redirection), the context can be limited to only 2 schemes: `javascript:` and `data:`. 
Replacing those 2 strings, and redirecting the malicious user on a `403 Forbidden` error page, makes harder exploit the highlighted vulnerability :).

Note: The changes don't implement a WAF/similar, but simply replaces the malicious `URI-schemes` in order to avoid that any attacker could utilize them.